% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonify.R
\name{harmonify}
\alias{harmonify}
\title{Performs codon harmonization with the CHARMING algorithm}
\usage{
harmonify(
  seq_path,
  numMutants = 3,
  windowSize = 10,
  measure = "MM",
  outName,
  codonTableOrigin,
  codonTableDestination,
  save_plots = TRUE,
  save_fasta = TRUE
)
}
\arguments{
\item{seq_path}{String. The file path to a sequence you would like to harmonize in FASTA format.}

\item{numMutants}{Numeric. The desired number of synonymous harmonized mutants. Must be a integer greater than 1 (default = 3).}

\item{windowSize}{Numeric. The size of the sliding window to use with the desired codon usage measure. This value must be a positive integer greater than 1 (default = 10).}

\item{measure}{String. The desired codon calculator, must be either "MM" (for percent MinMax) or "GM" (for geometric mean). Defaults to "MM".}

\item{outName}{String. The name (without spaces) you wishes to associate with the harmonization.}

\item{codonTableOrigin}{String. The path to the codon usage table file for the origin species.}

\item{codonTableDestination}{String. The path to the codon usage table file for the destination species.}

\item{save_plots}{Logical. Generate plots.}

\item{save_fasta}{Logical. Generate FASTA files containing the harmonized mutants.}
}
\value{
A list containing:

\code{[[1]]}
A matrix with the following structure:

\describe{
  \item{\code{[,1]}}{Numeric. Represents the calculated distances or metrics.}
  \item{\code{[,2]}}{Character. Contains the codon sequences corresponding to the rows.}
  \item{\code{[,3]}}{Numeric vector. Represents a list of metrics or values associated with each codon.}
}
Each row corresponds to a harmonized sequence, containing:
\enumerate{
  \item The distance (Column 1).
  \item The sequence (Column 2).
  \item Associated metrics (Column 3).
}

\code{[[2]]}
A tibble data frame in wide format.

This tibble includes:
\itemize{
  \item The codon positions of the input sequence.
  \item The codon usage metrics for the sequence in the source organism.
  \item The codon usage metrics of the same sequence in the destination organism.
  \item The codon usage metrics for the harmonized sequences requested for the destination organism.
}
}
\description{
This function is a re-implementation in R of the Python script CHARMIMG.py (https://github.com/wrightgs/CHARMING) that implements the codon harmonization algorithm CHARMING (Codon HARMonizING) described in Wright et al. (2022).
}
\examples{
# Example: Perform codon harmonization using CHARMING
seq_path <- system.file("extdata", "MAD3_Scer.txt", package = "evobioqR")
codonTableOrigin <- system.file("extdata", "ScerCUB.txt", package = "evobioqR")
codonTableDestination <- system.file("extdata", "EcolCUB.txt", package = "evobioqR")
harmonify(seq_path = seq_path,
          numMutants = 2,
          windowSize = 17,
          measure = "MM",
          outName = "test",
          codonTableOrigin = codonTableOrigin,
          codonTableDestination = codonTableDestination,
          save_plots = FALSE,
          save_fasta = FALSE)

}
\references{
Wright, G., Rodriguez, A., Li, J., Milenkovic, T., Emrich, S. J., & Clark,
P. L. (2022). CHARMING: Harmonizing synonymous codon usage to replicate a
desired codon usage pattern. \emph{Protein Science}, \bold{31(1)}, 221-231.
}
