% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molR.R
\name{molR}
\alias{molR}
\title{molR}
\usage{
molR(
  file = NULL,
  backgroundColor = NULL,
  selectionStyles = NULL,
  overrideRepresentation = NULL,
  addRepresentation = NULL,
  quickStyles = NULL,
  width = 600,
  height = 400
)
}
\arguments{
\item{file}{The file path to a PDB or MolViewSpec (.mvsj) file.}

\item{backgroundColor}{A string specifying the background color.
Supported color names include any of the X11 colors (https://www.w3.org/TR/css-color-3/#svg-color).}

\item{selectionStyles}{A list specifying selections with their representation type and/or color, e.g.,
`list(list(selection = "A 12-38", representation = "spacefill", color = "red"), list(selection = "B 50-99", representation = "cartoon", color = "blue"))`.}

\item{overrideRepresentation}{A string vector specifying the representation type and the color theme of the polymer to add.
Supported representations include: "backbone", "ball-and-stick", "cartoon", "gaussian-surface",
"gaussian-volume", "interactions", "label", "line", "molecular-surface", "orientation", "point",
"putty", "spacefill".
Supported color themes include: "atom-id", "element-index", "element-symbol", "formal-charge",
"occupancy", "uncertainty", "chain-id", "chain-instance", "entity-id", "entity-source",
"model-index", "polymer-chain-id", "polymer-chain-instance", "structure-index", "cartoon",
"external-volume", "illustrative", "illustrative-type", "uniform", "accessible-surface-area",
"hydrophobicity", "molecule-type", "residue-name", "secondary-structure", "sequence-id".}

\item{addRepresentation}{A string vector specifying the additional representation type with color theme of the polymer to add.}

\item{quickStyles}{A string specifying the quick style to apply.
Supported options are "illustrative" and "stylize current".}

\item{width}{Width of the widget.}

\item{height}{Height of the widget.}
}
\value{
An HTML widget displaying the molecular structure.
}
\description{
Visualize molecular structures using a custom Mol* viewer embedded in R Markdown.
}
\references{
Sehnal, D., et al. (2021). Mol* Viewer: modern web app for 3D visualization and analysis of large biomolecular structures.
Nucleic Acids Research, 49(W1), W431–W437.

Midlik, A., et al. (2025). MolViewSpec: a Mol* extension for describing and sharing molecular visualizations.
Nucleic Acids Research, gkaf370.
}
