% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolstarVieweR.R
\name{MolstarVieweR}
\alias{MolstarVieweR}
\title{MolstarVieweR}
\usage{
MolstarVieweR(
  pdbid = NULL,
  pdbData = NULL,
  pdbUrl = NULL,
  width = "100\%",
  height = "600px"
)
}
\arguments{
\item{pdbid}{A string containing a valid PDB identifier}

\item{pdbData}{A string containing PDB data}

\item{pdbUrl}{A URL to a PDB file}

\item{width}{Width of the widget}

\item{height}{Height of the widget}
}
\value{
An HTML widget displaying the molecular structure if a valid PDB identifier, URL, or data is provided. Otherwise, only launches the Mol* viewer widget.
}
\description{
Visualize molecular structures using the Mol* viewer
}
\references{
Sehnal, D., Bittrich, S., Deshpande, M., Svobodová, R., Berka, K., Bazgier, V., ... & Rose, A. S. (2021).
Mol* Viewer: modern web app for 3D visualization and analysis of large biomolecular structures.
\emph{Nucleic Acids Research}, \bold{49(W1)}, W431–W437.
}
