% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saltify.R
\name{saltify}
\alias{saltify}
\title{Ammonium Sulfate Saturation Calculator}
\usage{
saltify(initial_volume, desired_percentage, starting_percentage, temperature)
}
\arguments{
\item{initial_volume}{Numeric. The initial volume of the solution in milliliters (mL).}

\item{desired_percentage}{Numeric. The desired percentage of ammonium sulfate saturation.}

\item{starting_percentage}{Numeric. The starting percentage of ammonium sulfate saturation.}

\item{temperature}{Numeric. The temperature in degrees Celsius (째C) at which the calculation should be made.}
}
\value{
A string. A message informing you of how many grams of ammonium sulfate
        should be added to achieve the desired saturation at the specified temperature.
}
\description{
This function calculates the amount in grams of ammonium sulfate to be added to
a solution to achieve a desired percentage of saturation at a specified temperature.
It uses a linear model to interpolate ammonium sulfate properties (specific volume,
saturation, and molarity) between discrete temperatures (0째C to 25째C).
}
\details{
\deqn{
\mathrm{weight (g)} = \frac{weight_{Sat} \times (M_2 - M_1)}{M_{Sat} - \left(\frac{\nu}{1000} \times 132.14 \times M_{Sat} \times M_2\right)}
}

Where:
\itemize{
  \item \eqn{weight_{Sat}}: The number of grams per liter in a saturated solution of ammonium sulfate.
  \item \eqn{M_2}: The molarity we want to achieve.
  \item \eqn{M_1}: The starting molarity.
  \item \eqn{M_{Sat}}: The molarity of a saturated solution of ammonium sulfate.
  \item \eqn{\nu}: The specific volume of ammonium sulfate.
  \item \eqn{\mathrm{(NH_4)_2SO_4}}, 132.14 g/mol
}

The function uses a linear model to interpolate values of \eqn{weight_{Sat}}, \eqn{M_{Sat}}, and \eqn{\nu}
based on temperature, as these properties vary with temperature. The user provides the initial
volume, desired saturation percentage, starting saturation percentage, and temperature.
}
\examples{
# Example: Calculate the amount of ammonium sulfate for a 100 mL solution
# with a starting saturation of 30\%, a desired saturation of 50\%, and a temperature of 20째C.
saltify(100, 50, 30, 20)

}
\references{
Wingfield, P. (1998). Protein precipitation using ammonium sulfate.
\emph{Current protocols in protein science}, \bold{13(1)}, A-3F.
}
